+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Include comments in the line length in `TooLongLineRule`.
 - Update severity metadata for all rules.
 
 ### Fixed